fix(scan-code): local install

+++ packages/ScanCode/scripts/postinstall.js
@@ -7,6 +7,8 @@ const depsRoot = path.resolve(packageRoot, '../..')
 
 const wasm = path.resolve(depsRoot, 'zxing-wasm/dist/full/zxing_full.wasm')
 const target = path.resolve(process.env.INIT_CWD, 'public/zxing_full.wasm')
-fs.copyFileSync(wasm, target)
+if (fs.existsSync(wasm)) {
+  fs.copyFileSync(wasm, target)
+}
 
 // throw new Error(JSON.stringify(process.env, null, 2))
